+++ cli/floatingip.go
@@ -23,6 +23,7 @@ func newFloatingIPCommand(cli *CLI) *cobra.Command {
 		newFloatingIPDisableProtectionCommand(cli),
 		newFloatingIPAddLabelCommand(cli),
 		newFloatingIPRemoveLabelCommand(cli),
+		newFloatingIPSetRDNSCommand(cli),
 	)
 	return cmd
 }